Fix for clippy

+++ crates/vizia_winit/src/window.rs
@@ -554,7 +554,7 @@ impl View for Window {
 
             WindowEvent::SetImeCursorArea(position, size) => {
                 let position = PhysicalPosition::new(position.0 as i32, position.1 as i32);
-                let size = PhysicalSize::new(size.0 as u32, size.1 as u32);
+                let size = PhysicalSize::new(size.0, size.1);
                 self.window().set_ime_cursor_area(position, size);
             }
